[develop] Adjusting memoization v1 and v2
