+++ other_topics/memoization/v1/utils/memoize.py
@@ -1,0 +1,10 @@
+def memoize(f):
+    memo = {}
+
+    def helper(x):
+        if x not in memo:
+            print(f'Return from function {f.__name__}() not in cache, executing now with args [{x}]')
+            memo[x] = f(x)
+        return memo[x]
+
+    return helper
