+++ other_topics/memoization/v2/utils/memoize.py
@@ -6,5 +6,6 @@ class Memoize:
 
     def __call__(self, *args):
         if args not in self.memo:
+            print(f'Return from function {self.fn.__name__}() not in cache, executing now with args {list(args)}')
             self.memo[args] = self.fn(*args)
         return self.memo[args]
